requirements.txt
bluepy==1.1.4
git+https://github.com/CY8s/pycirculate.git#egg=pycirculate
paho-mqtt==1.3.1
